services/meeting/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic-settings
python-jose[cryptography]
cryptography
websockets
httpx
minio
redis
celery
pgvector
PyJWT
